+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/CalibrateSensorActivity.java
@@ -22,7 +22,6 @@ import android.content.DialogInterface;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
@@ -37,51 +36,20 @@ import com.ftdi.j2xx.FT_Device;
 import org.akvo.caddisfly.R;
 import org.akvo.caddisfly.app.MainApp;
 import org.akvo.caddisfly.util.AlertUtils;
-import org.akvo.caddisfly.util.ApiUtils;
 
-public class CalibrateSensorActivity extends AppCompatActivity {
-
-    //private static final int readLength = 512;
-    //    private static final String ACTION_USB_PERMISSION = "org.akvo.caddisfly.USB_PERMISSION";
-//    private static final int DEFAULT_BAUD_RATE = 9600;
-//    private static final int DEFAULT_BUFFER_SIZE = 1028;
-//    private static final int REQUEST_DELAY = 4000;
-    //private static final int INITIAL_DELAY = 1000;
-    // original ///////////////////////////////
-    //private final StringBuilder mReadData = new StringBuilder();
-    //public readThread read_thread;
-    //Toast debugToast;
-    /*graphical objects*/
-    //ArrayAdapter<CharSequence> portAdapter;
-    //ArrayList<CharSequence> portNumberList;
-    //private boolean bReadThreadGoing = false;
+public class CalibrateSensorActivity extends BaseActivity {
 
     private D2xxManager ftManager;
     private FT_Device ftDev = null;
     private int DevCount = -1;
     private int currentIndex = -1;
-    /*local variables*/
+
     private int baudRate; /*baud rate*/
     private byte stopBit; /*1:1stop bits, 2:2 stop bits*/
     private byte dataBit; /*8:8bit, 7: 7bit*/
     private byte parity;  /* 0: none, 1: odd, 2: even, 3: mark, 4: space*/
     private byte flowControl; /*0:none, 1: flow control(CTS,RTS)*/
-    //private int portNumber; /*port number*/
-//    private byte[] readData;
-//    private char[] readDataToText;
-    //private String mEc25Value = "";
-    //private String mTemperature = "";
     private Context mContext;
-    //    final Handler handler = new Handler() {
-//        @Override
-//        public void handleMessage(Message msg) {
-//            if (byteCount > 0) {
-//                Toast.makeText(mContext, String.copyValueOf(readDataToText, 0, byteCount), Toast.LENGTH_LONG).show();
-//            }
-//        }
-//    };
-    //private String mResult = "";
-    //private int delay = INITIAL_DELAY;
 
     @Override
     public void onStart() {
@@ -99,14 +67,6 @@ public class CalibrateSensorActivity extends AppCompatActivity {
     public void onDestroy() {
         super.onDestroy();
     }
-
-//    public void notifyUSBDeviceAttach() {
-//        createDeviceList();
-//    }
-//
-//    public void notifyUSBDeviceDetach() {
-//        disconnect();
-//    }
 
     private void createDeviceList() {
         int tempDevCount = ftManager.createDeviceInfoList(this);
@@ -164,17 +124,7 @@ public class CalibrateSensorActivity extends AppCompatActivity {
 
         if (ftDev.isOpen()) {
             currentIndex = openIndex;
-
-//            if (!bReadThreadGoing) {
-//                read_thread = new readThread(handler);
-//                read_thread.start();
-//                bReadThreadGoing = true;
-//            }
         }
-        //else {
-        //Toast.makeText(this, "Need to get permission!", Toast.LENGTH_SHORT).show();
-        //}
-
         SetConfig(baudRate, dataBit, stopBit, parity, flowControl);
 
     }
@@ -309,8 +259,6 @@ public class CalibrateSensorActivity extends AppCompatActivity {
             ex.printStackTrace();
         }
 
-        ApiUtils.lockScreenOrientation(this);
-
         final ViewAnimator viewAnimator = (ViewAnimator) findViewById(R.id.viewAnimator);
 
         Button nextButton = (Button) findViewById(R.id.startButton);
@@ -326,17 +274,6 @@ public class CalibrateSensorActivity extends AppCompatActivity {
                     mainApp.currentTestInfo.getName(conf.locale.getLanguage()));
         }
 
-
-        //http://developer.android.com/guide/topics/connectivity/usb/host.html
-//        IntentFilter filter = new IntentFilter();
-//        filter.addAction(UsbManager.ACTION_USB_DEVICE_ATTACHED);
-//        filter.addAction(UsbManager.ACTION_USB_DEVICE_DETACHED);
-//        registerReceiver(mUsbReceiver, filter);
-
-        //readData = new byte[readLength];
-        //readDataToText = new char[readLength];
-
-        /* by default it is 9600 */
         baudRate = 9600;
 
         stopBit = 1;
@@ -346,9 +283,6 @@ public class CalibrateSensorActivity extends AppCompatActivity {
         parity = 0;
 
         flowControl = 0;
-
-        //portNumber = 1;
-
 
         nextButton.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -448,7 +382,6 @@ public class CalibrateSensorActivity extends AppCompatActivity {
                 }
             }
         });
-
     }
 
     private boolean validInput(String input) {
@@ -460,43 +393,4 @@ public class CalibrateSensorActivity extends AppCompatActivity {
                 Context.INPUT_METHOD_SERVICE);
         imm.hideSoftInputFromWindow(input.getWindowToken(), 0);
     }
-
-//    private class readThread extends Thread {
-//        Handler mHandler;
-//
-//        readThread(Handler h) {
-//            mHandler = h;
-//            this.setPriority(Thread.MIN_PRIORITY);
-//        }
-//
-//        @Override
-//        public void run() {
-//            int i;
-//
-//            while (bReadThreadGoing) {
-//                try {
-//                    Thread.sleep(50);
-//                } catch (InterruptedException ignored) {
-//                }
-//
-//                synchronized (ftDev) {
-//                    int byteCount = ftDev.getQueueStatus();
-//                    if (byteCount > 0) {
-//
-//                        if (byteCount > readLength) {
-//                            byteCount = readLength;
-//                        }
-//
-//                        ftDev.read(readData, byteCount);
-//                        for (i = 0; i < byteCount; i++) {
-//                            readDataToText[i] = (char) readData[i];
-//                        }
-//                        Message msg = mHandler.obtainMessage();
-//                        mHandler.sendMessage(msg);
-//                    }
-//                }
-//            }
-//        }
-//    }
-
 }