+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/SwatchActivity.java
@@ -17,18 +17,16 @@
 package org.akvo.caddisfly.ui;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 
 import org.akvo.caddisfly.R;
-import org.akvo.caddisfly.util.ApiUtils;
 
-public class SwatchActivity extends AppCompatActivity {
+public class SwatchActivity extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_swatch);
-        ApiUtils.lockScreenOrientation(this);
+
         if (savedInstanceState == null) {
 
             SwatchFragment fragment = new SwatchFragment();